fix(lints): clippy 1.96 verde — el runner CI adopto rustc 1.96.1 (local iba en 1.95) y -D warnings aflora 10 lints nuevos en codigo preexistente (sort_by_key, if-let, useless vec!, type alias, doc-indent, cast f32, import duplicado)

+++ control-center/src-tauri/src/memory/doctor.rs
@@ -303,7 +303,7 @@ fn check_evals() -> DoctorCheck {
     evals_verdict(
         golden.aggregate.recall_at_k as f32,
         golden.degraded,
-        smoke.recall_at_k as f32,
+        smoke.recall_at_k,
         smoke.secret_leak_count,
         smoke.stale_leak_count,
     )
